+++ internal/scanner/languages.go
@@ -1,5 +1,5 @@
 // Copyright 2024 Google LLC
-// Copyright 2025 Ian Lewis
+// Copyright 2025 Ian Lewis, Marcin Wiśniowski
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -647,6 +647,14 @@ var LanguagesConfig = map[string]*Config{
 		LineComments:      hashLineComments,
 		MultilineComments: nil,
 		Strings:           cStrings,
+	},
+	"Svelte": {
+		LineComments: cLineComments,
+		MultilineComments: concat(
+			xmlBlockComments,
+			cBlockComments,
+		),
+		Strings: cStrings,
 	},
 	"Swift": {
 		LineComments:      cLineComments,